Backend/Ransom-ware/requirements.txt
cryptography
python-dotenv
Pillow
opencv-python
pygetwindow
pyautogui
pynput
tkinter
PyInstaller
magic
subprocess
ffpyplayer
pygam